refactor to improve type checking for users
