+++ cubewalkers/__init__.py
@@ -6,5 +6,15 @@ from cubewalkers import (
     simulation,
     update_schemes,
 )
-from cubewalkers.Experiment import Experiment
-from cubewalkers.Model import Model
+from cubewalkers._experiment import Experiment as Experiment
+from cubewalkers._model import Model as Model
+
+__all__ = [
+    conversions,
+    initial_conditions,
+    parser,
+    simulation,
+    update_schemes,
+    Experiment,
+    Model,
+]
