+++ tests/__init__.py
@@ -1,3 +1,20 @@
-from cubewalkers import conversions, parser, simulation, update_schemes, initial_conditions
-from cubewalkers.Model import Model
-from cubewalkers.Experiment import Experiment
+# type: ignore
+from cubewalkers import (
+    conversions,
+    initial_conditions,
+    parser,
+    simulation,
+    update_schemes,
+)
+from cubewalkers._experiment import Experiment as Experiment
+from cubewalkers._model import Model as Model
+
+__all__ = [
+    conversions,
+    initial_conditions,
+    parser,
+    simulation,
+    update_schemes,
+    Experiment,
+    Model,
+]
